requirements.txt
Django==4.2.7
Pillow==10.1.0
python-magic==0.4.27
gunicorn==21.2.0
whitenoise==6.6.0
